fix mypy issues [mypy]
