+++ deepdoctection/utils/fs.py
@@ -58,7 +58,7 @@ def sizeof_fmt(num: float, suffix: str = "B") -> str:
 
 # Copyright (c) Tensorpack Contributors
 # Licensed under the Apache License, Version 2.0 (the "License")
-def mkdir_p(dir_name: str) -> None:
+def mkdir_p(dir_name: Pathlike) -> None:
     """
     Like "mkdir -p", make a dir recursively, but do nothing if the dir exists
 
@@ -76,7 +76,7 @@ def mkdir_p(dir_name: str) -> None:
 
 # Copyright (c) Tensorpack Contributors
 # Licensed under the Apache License, Version 2.0 (the "License")
-def download(url: str, directory: str, file_name: Optional[str] = None, expect_size: Optional[int] = None) -> str:
+def download(url: str, directory: Pathlike, file_name: Optional[str] = None, expect_size: Optional[int] = None) -> str:
     """
     Download URL to a directory. Will figure out the filename automatically from URL, if not given.
     """
@@ -122,7 +122,7 @@ def download(url: str, directory: str, file_name: Optional[str] = None, expect_s
     return f_path
 
 
-def is_file_extension(file_name: str, extension: Union[str, Sequence[str]]) -> bool:
+def is_file_extension(file_name: Pathlike, extension: Union[str, Sequence[str]]) -> bool:
     """
     Check if a given file name has a given extension
 
@@ -133,23 +133,6 @@ def is_file_extension(file_name: str, extension: Union[str, Sequence[str]]) -> b
     if isinstance(extension, str):
         return os.path.splitext(file_name)[-1].lower() == extension
     return os.path.splitext(file_name)[-1].lower() in extension
-
-
-class LoadImageFunc(Protocol):  # pylint: disable = R0903
-    """
-    Protocol for typing load_image_from_file
-    """
-
-    @overload
-    def __call__(self, path: str, type_id: Literal["np"]) -> Optional[ImageType]:
-        ...
-
-    @overload
-    def __call__(self, path: str, type_id: Literal["b64"]) -> Optional[str]:
-        ...
-
-    def __call__(self, path: str, type_id: Literal["np", "b64"]) -> Optional[Union[str, ImageType]]:
-        ...
 
 
 @overload
@@ -190,7 +173,7 @@ def load_image_from_file(path: Pathlike, type_id: Literal["np", "b64"] = "np") -
     return image
 
 
-def load_bytes_from_pdf_file(path: str) -> bytes:
+def load_bytes_from_pdf_file(path: Pathlike) -> bytes:
     """
     Loads a pdf file with one single page and passes back a bytes' representation of this file. Can be converted into
     a numpy or directly passed to the attr: image of Image.
@@ -209,9 +192,26 @@ def load_bytes_from_pdf_file(path: str) -> bytes:
     return buffer.getvalue()
 
 
+class LoadImageFunc(Protocol):  # pylint: disable = R0903
+    """
+    Protocol for typing load_image_from_file
+    """
+
+    @overload
+    def __call__(self, path: Pathlike, type_id: Literal["np"]) -> Optional[ImageType]:
+        ...
+
+    @overload
+    def __call__(self, path: Pathlike, type_id: Literal["b64"]) -> Optional[str]:
+        ...
+
+    def __call__(self, path: Pathlike, type_id: Literal["np", "b64"]) -> Optional[Union[str, ImageType]]:
+        ...
+
+
 def get_load_image_func(
-    path: str,
-) -> Union[LoadImageFunc, Callable[[str], bytes]]:
+    path: Pathlike,
+) -> Union[LoadImageFunc, Callable[[Pathlike], bytes]]:
     """
     Return the loading function according to its file extension.
